Initial Commit
Kanawha is a hobby kernel I've been working on in my free time for the past few weeks,
I started it last spring but took a few months off at the start of this summer.

This is primarily a "sandbox" kernel for me to play around with ideas I have, and to
try out writing some core kernel systems / the user-space to kernel API.

I have no idea how far I'm going to take the kernel but I'm having fun so far,
so I'll keep going as long as I feel like.