environ Syscall
Because kanawha is trying to avoid making any assumptions about the
memory layout of a user process (no automatic stack or brk), then
we run into trouble passing information across an exec system call,
which will wipe the mmap state fully currently. So passing ARGV or other
info like which file descriptor is STDIN/STDOUT/STDERR (0 1 and 2 won't work
with the current system) isn't really possible.
We could do it through files, but I'm not a huge fan of that idea, and
we ultimately run into the same problem as with the STD stream file descriptors.

So instead, I've added a simple key-value store for environment variables, managed
for each process by the kernel. Each process will inherit it's environment from
it's parent, and so we can define variables such as "ARGV", "STDIN", "STDOUT", "STDERR"
to pass data across an "exec" syscall.

Right now processes are allowed to allocate potentially unlimited kernel memory
just by allocating more and more environment variables (currently we actually
have the same problem for mmap regions) so this will need to be managed when
we eventually introduce some form of resource limits.