x64 IO/APIC Support
Can parse the MADT for IO/APIC's and register them as IRQ devices.

We also are registering a single domain for the PIC, even though it is inactive,
so that we can link from the IO/APIC to this domain based on MADT Override entries.

Then legacy device drivers like the PS/2 and Serial Ports can attach handlers
to the PIC/ISA IRQ Domain at the legacy defined IRQ's.