More File Manipulation Syscalls
Added the seek, attr, and childname syscalls.

seek is used for moving around a file descriptor's offset,
and in accordance, the read and write syscalls have been modified
to read and write from the seek offset, instead of taking in an
offset as an argument.

I was trying to avoid the need for seek, but I'm coming to realize
that seek makes stream-like files much much simpler to deal with,
and overall is ridiculously simple to implement.

attr is a generic "file-info" syscall (think of stat on linux), which
currently just enables accessing the file size, current seek offset (tell)
and the number of children of the file (for directories).

Then given a child index in [0,FILE_ATTR_CHILDREN) the childname syscall
gives a way to get the file-name of the child.

More needs to be done for protections still though,
and directories need to be protected with some permissions.
Ideally in addition to FILE_PERM_READ and FILE_PERM_WRITE, I'd
like there to be a FILE_PERM_DIR_READ and FILE_PERM_DIR_WRITE,
to stop users from reading a directory or creating/destroying files
in the directory, but it really depends on how well different file
systems would support that.