Process Spawning
Introduces spawn, reap, and getpid system calls.

Theres a lot in this one.

This all still needs to be documented, but I'm putting it off,
mostly out of laziness, but partially because the API's keep changing.
And I'd like to avoid having to redo documentation every time
I make a change to a system call.

So for now I'm going to leave things vague. (I'm certain there are
still deadlock situations and other general bugs in this commit)