Improved x64 Thread Volatility Testing
I added the x64 thread volatility test in the previous commit and
didn't really explain.

All it does, is spawn a background thread that sets every GPR to a known value,
and then spins with interrupts on to make sure those values don't change,
this way, if an interrupt or interrupt driven thread switch accidentally corrupts a register,
this thread will notice, and panic, which can be really useful when messing around with
the (somewhat messy) assembly transitions during interrupts, exceptions, and thread switches.