Reworks/Fixes for -O0
Some things were simple fixes (moving the kernel base to have
enough room for the much larger binary)

Others are trickier, and are still issues on -O3. A big one is setting
all trap gates to interrupt gates, even for exception vectors on x64.

When a user exception occurs (usually a page fault), then there is a race
condition between the entry to the trap and the "swapgs" instruction, if another
interrupt occurs in the meantime, then the gs register won't be swapped correctly.

So by marking all IDT entries as "interrupt gates" instead of "trap gates" we can
be sure that IF will be disabled on entry, avoiding the race.