Loading Root FS and Fixing x64 Stack Overflow Issue
Re-added support for loading a basic rootfs during boot to load
files from. I realized it's better to leave the job of loading modules from that
rootfs to the "init" process instead of trying to support that inside the kernel though.

I also fixed a signficant issue with the x64 interrupt driven thread switching.
It's really just a single line fix in "arch/x64/exception.c" which resets the
stored kernel stack RSP value if we came from an outer ring. If we didn't
do this, then the RSP value would creep upwards from interrupt driven thread
switches and we'd run into a stack overflow.