Even More Filesystem Changes (Directory Support)
This contains some major changes to the filesystem API,
making the whole thing much more similar in design to the
Linux VFS than it was previously.

And this makes test programs like "cat" and "ls" possible to
implement in Elk, and should hopefully make implementing a "real"
filesystem (a stub for ext2 has been started) much easier.