Memory Flag Locking and Compaction
Previously we weren't locking the memory flags
structures, and over time they would grow larger
than necessary (adjacent regions with the same flags
weren't being merged).

But now that we are allocating virtual stacks for each thread on x64,
this means the flags are being updated even after the serial
boot phase, and if the map became too fragmented, we could fail
to allocate a new thread, even if there was plenty of available
physical and virtual memory.