MMAP_EXACT and mmap hinting
Changed the default behaviour of "mmap" to treat
the provided address (now a pointer) as a suggestion
which can be modified by the kernel to move the mapping as
is appropriate.

To get the old semantics, with "hint_offset" not being a suggestion,
we can pass MMAP_EXACT in the "mmap_flags" field.