Removing mkfifo system call
Right now, the mkfifo syscall is no where near implementable, and
I'm not even sure that I want to have filesystem named pipes.
Instead I'm going to implement anonymous pipes for the short term,
and come back to the issue of named pipes when it will inevitably come
up again.