pipe Syscall and Removing NULL_FD
Added the "pipe" syscall to generate anonymous pipes which can
be used for IPC (at least between related processes).

In the same motion, removed NULL_FD as a concept, which makes sense
since "open" returns an "int" instead of an "fd_t", and we would
like fd_t zero to be available for use as stdin.