insmod and rmmod Syscalls
Added the insmod and rmmod system calls, still in an
early stage, and VERY subject to change.

But this gives a simple way to start loading code and data into
the kernel at runtime. "memdump.c" can be compiled to a module
with when loaded just dumps the amount of free memory in the kernel page
allocators. (This is useful for testing)