EXT2: Extending Directories
If we find an EXT2 directory which is too small, then we now can extend
the linked list inside the directory in order to add more files to the
directory.

This means we can add files to directories we create with mkdir, and
that we can create the "." and ".." directory entries on a "mkdir".