PCI MSI/MSI-X Probing
Probe all PCI devices during boot for MSI and MSI-X capability
structures, and start making IRQ's available to PCI drivers.

Right now this doesn't do much other than detect MSI/MSI-X but I'd like
for a PCI device driver to be able to completely ignore what interrupt
method it's actually using (INT-X, MSI, MSI-X) and just interface with
PCI "IRQ's".

This may not actually be possible, and some drivers might need to behave
differently depending on which IRQ "mode" is used, but I'm going to run
with this until it breaks for now.