MSI and MSI-X Support
Only MSI has really been tested so far (using the QEMU EDU device),
but the logic is there to allow both MSI and MSI-X interrupts to be
used by PCI drivers in the future.

This also includes support for "PCI Mailboxes" which are essentially
just devices which can accept MSI messages, and this gives a way for
those devices to specify the format of those messages. (Currently all
that is supported is the x64 LAPIC, and it doesn't do a very good job to
say the least).