Virtio Probing
Start probing for Virtio devices and building up a subsystem about
managing virtio devices.

Currently trying to seperate the virtio drivers themeselves from the
"transport" which is underneath them, so that the same virtio driver
can potentially work for both pure MMIO virtio and PCI based virtio.

For testing I started writing a virtio-console driver, but it's far from
finished, and the PCI transport itself is full of unimplemented
functions.