Refactor: Put proc Files into subdirectory
Move a number of "process" focuses files into a specific
subdirectory. This helps clean up a bit, especially for the top
level "mmap.c/h" and "file.c/h" which really refer to the
"mmap" address space of a specific process, and the "file(_table)"
of a process, not the mmap syscall or files in general.