Framework for mclose and mcreate syscalls
One of the kind of dumb, but could be interesting, things I want to do
is expose the address space abstraction more directly to userspace.
By allowing processes to create, modify, and enable multiple "aspaces"
at once.

This means "mmap" needs to take another argument, the address space
"descriptor" to modify (similar to file descriptors ideally) or
*_CURRENT to modify the active aspace.

In the end I want this change to make a couple of other complicated
syscalls able to be implemented as userspace libraries, instead of in
the kernel itself. The most notable being "exec", (hopefully also making
the whole "environ" syscall unnecessary as well, because originally that
was just a hack I put in place to avoid assuming a stack).