Documentation with Sphinx
I'm not sure that I will ever actually write much documentation for this
project. But I think the likelihood that I document API's when I'm
writing them will go up if it looks slightly pretty when I do it, and if
the framework is already in place to make it easy.