rr_sched Bug Fix
I had forgotten to clear the "current_rr_thread" of a CPU when
scheduling the idle thread.

So in a situation where we start running IDLE from thread N,
and then thread N becomes the only thread available
to run, the scheduler would simply never run thread N because it
considered it to be the current thread. Should be fixed now,
but regardless I need to sit down and write a better scheduler than the
pretty terrible "rr_sched" I wrote just to get things up and running in
the short term.