Timer Framework Rework
I realized that the old timer system was really just working
because of the fact that the only event we ever consistently enqueued
was the "rr_sched_kick" callback. And so the old "queue" system really
easily deadlocked when multiple sources were placing events into the
queue.

Maintaining the queue with relative times just doesn't really work when
you don't know how long into the "first" event's timer callback you are
enqueueing. So every callback would have a deviation somewhere between 0
and the "wait_for" of the first event in the queue. (This was bad).

Instead I'm going to run with a system where N timers are provided,
and these can be "reserved" for use by a single part of the kernel.

To make this work for larger systems, we then place the "periodic"
subsystem which grabs exactly 1 timer and maintains a basic tick rate for the
system at all times, and can be tracking a number of different
"periodic" events accurately, assuming that these periodic events have a
period which is a multiple of the tick rate (10ms by default).