Ignore Unknown Modifiers in doprnt + puts newline
If an unrecognized modifier is seen during a doprnt, do not return from
handling the escape sequence, and instead continue, ignoring the unknown
character.

puts needs to output a newline after printing the string (the compiler
inserts a number of fputs/puts to replace printf's when a format string
has no arguments.