RISCV Basic Booting
Need to change how the "identity map" is setup in the kernel to support
RISC-V because RISC-V does not let us specify exactly where in physical
memory the kernel will be loaded.
So RISC-V maps the kernel into virtual memory seperately from how the
kernel will allow access to an identity map in high-mem for accessing
physical memory.