Questionable PCI MMIO BAR Allocation
When running on RISCV without a bootloader, the PCI BAR's may not
actually be allocated correctly. To try and get around this, I've added
an option which will allocate some portion of RAM and manually map any
MMIO bars it finds unmapped on top of the RAM (effectively shadowing the
actual memory underneath). The "real" solution here is just to install a
bootloader on whatever system which can do the PCI mapping for the
kernel. But this very hackish solution sometimes works...