virtio_transact and virtio request polling
Added the virtio_transact helper function(s) which can handle a simple
virtio transaction (DMA allocation, copying structs, launching the
request, awaiting the request, and copying the result back) which
basically all virtio drivers will need to perform in the same order.

Also changed the virtio_request_await function so that instead of
sitting in a loop waiting for an interrupt to occur, the loop will
actually poll the device. (This might speed up some short requests, and
at times the kernel was freezing due to lost used notifications from the device which
were never re-sent)