Simple virtio_gpu FB Driver (BUGGED)
The driver can actually display a simple terminal for now.
This version of the driver will stop changing the display properly
after a certain number of lines is printed out in the usermode
terminal though. And I suspect this driver is leaving interrupts
disabled somewhere because the timer stops ticking as well.