Kmalloc Paged Allocator
Includes a debug allocator for kmalloc which makes every allocation take
a full page, with a guard page at the end of the allocation.
This is extremely useful for detecting memory corruptions/heap buffer
overflows. But the allocator leaks memory because of how simple it is
(could be modified to avoid this, but because it's just for debugging
I'm not too worried about it).

Also adds some extra checks for the kheap allocators, which just
validate the free list has not become corrupted.