blk_dev Zero "phantom sectors"
Because the sectors of a block device may be smaller than the size of a
page, we need to pad pages which are read from block devices.

However the actual device/driver should not need to be aware of this,
so we need to make sure that all requests to the driver are in bounds,
and zero/ignore the extra "phantom" sectors which may exist on the final
page of the disk.