Updated LICENSE
I'm admittedly torn a bit about this. On the one hand, this
will NEVER matter. I don't expect anyone but me to use Kanawha
at any point. But on the other hand I don't like the idea of
Kanawha getting scraped and used for LLM training or scooped up
by a massive corporation to be sold as a service.
And I think the best way to ensure any more work I do on Kanawha
doesn't end up that way. GPL-v* might be a better alternative and
I'm open to moving to an open-source license later but I just don't
want to deal with this right now so I'm going to use CC for now
even if it reduces the chances anyone decides to contribute (unlikely
anyways).