ACPI AML Parsing (WIP)
I decided I didn't want to use ACPIA.
This was a bad decision.

AML wouldn't be that awful to parse, it's just the MethodInvocations
assume that you know how many arguments each method takes.
So you must concurrently *run* the AML program as you parse it,
which is insane.

why couldn't they just add a "number of arguments" to the invocation...