Refactor Registering PCI Segments/Probing
Split PCI Configuration Access Mechanisms from
any specific segment, and act as if any provided CAM could access any
segment. This is not efficient, but it is about as flexible as possible
(If one CAM fails to access a configuration space location, just try the
next one until one is found which works)

Ideally later on we will have a system to "order" the PCI access
mechanisms so a "better" (ECAM > CAM) mechanism is always chosen first
to help with performance.