Have pipes respect non-blocking read/writes
Previously pipes would simply reject any non-blocking reads or writes,
(this was pretty naive).

Just check if any bytes are available and if they are, the write is
non-blocking so we can return instantly.

This does not account for the locks on the pipes themselves, but no one
should ever block while holding a pipe lock so waiting for the pipe
locks should not be truly "blocking"