Massive Backlog (POTENTIAL REBASE/SPLIT LATER)
I'm primarily focusing on updating the actual kernel,
so ELK is changing constantly to adapt to new system calls
and such. As well as adding random new libc/libm/posix functions
I didn't have previously.

So this is just a massive number of changes which I can't be bothered to
split apart for a hobby project.

(I'm also slightly scared the laptop I'm working on is going to die
soon, and so I want to make this commit ASAP so I don't lose all the
changes if/when that happens).