Network Device Framework Progress
Instead of having whole "network" devices, I'd like to have one device
for different layers of the network stack.

This way we can still allow for the network stack to be built up in the
kernel as a set of drivers which build upon one another, but still
allow userspace to access the networking stack at any level (and
potentially write userlevel networking stacks more easily).