Generic Device Interfaces
I'm trying to reduce the duplication of the same "register/unregister"
functions for each device type.

So now it's ALL MACROS (THIS IS OBVIOUSLY A GOOD THING AND NOT GOING TO
BE A NIGHTMARE TO MAINTAIN IN THE FUTURE YOU HAVE NO PROOF I AM INSANE
AND I WILL WORSHIP THE XMACROS UNTIL THE END OF MY DAYS) :D

The bulk of every device interface is now the "sysfs" layer, which is
also a bunch of boilerplate for most devices, but I'm going to leave it
alone for the immediate moment.

I also haven't actually made every type of device use the new macros.
Currently it's just the "char_dev" framework which has been ported over.