Ethernet and IP + Registry
The dev.h macros turned out to be a bad-ish idea because it meant I
couldn't easily control extra fields inside the struct's of devices (and
add a generic layer on top of a device type (such as multiplexing on top
of ethernet receive callbacks)).

This also comes with a big step forward in terms of the ethernet and IP
frameworks.