Tasklets
Added "tasklets" which give a simple way to add "high" vs. "low" half
interrupt handlers where the actual interrupt handler just triggers
another function to be run later inside a kernel thread (and so not
inside the interrupt context)

This is important for interrupts which have a lot of work which needs to
be done/needs to be done outside of an interrupt, such as when a new
device is attached and drivers need to be enumerated to probe the
device. (*cough* *cough* the USB framework which is in progress)