USB XHCI Command Queues and Attach (IN PROGRESS)
More progress on the USB XHCI device driver.

I still don't have a complete picture on what API the specific device
drivers will need to be presented in order to drive a USB device, (and
hopefully abstract the host controller away) but I'm hoping that API
will come naturally out of writing the XHCI driver first.

This is also the first actual use of a tasklet!