Add "free" sysfs info node
Simple way to get the amount of free memory in the system as it is
running. This immediately seems to have pointed out that there is at
least 1 if not many memory leaks around the kernel.

So I'm going to take a closer look and see if I can patch the leaks by
adding more provenance/flags to the memory allocators in the kernel
(kmalloc needs the GFP flags like Linux has by now)

The code for adding the "free" node is in "klog" for now but eventually
that should be moved.