Move fs_node_ops and fs_file_ops into "backing"
Ultimately my goal is to full deattach the fs_node struct from any
containing structure so that we can have "dead" fs_nodes with no backing
without invalidating pointers to the fs_node structure itself.

To work towards that this commit moves the fs_node_ops and fs_file_ops
into an inner "backing" structure which contains the state inside an
fs_node which needs to be accessible.