ops.h Pre and Post
Add the ability to run PRE and POST statements on every method call.

This means we need to store the return value of every method call, and
indirectly means that methods cannot return void.
This is not a big deal, as functions which previously returned void
(there were only 2) will now return an int, which can also serve as an
errno value.