fs_mount load/unload node give entire fs_node
There was some thought that only giving the backing of an fs_node to the
laod/unload functions would allow hiding the contents of the main
fs_node structure from the larger kernel.

But it makes more sense to give drivers access so that they can track
the fs_node of all loaded files. (And possibly forcefully invalid the
structure if a node is destroyed by something which cannot be stopped,
such as a device disconnect)